refactor: features/*/pages → src/pages 라우팅 진입점 통합 마이그레이션
- 9개 Page 컴포넌트를 features/*/pages/ → src/pages/로 이동
- auth 페이지 5종(Login, Signup, Kakao/Naver/GoogleCallback) → src/pages/auth/
- ChatPage, NotesPage, StoragePage, PricingPage → src/pages/
- 이동된 파일의 상대경로 import를 @/ 절대경로로 일괄 수정
- routes.tsx import를 @/pages/... 단일 네임스페이스로 통일
- konva/perfect-freehand 캔버스 로직은 features/editor에 그대로 유지
