+++ src/pages/auth/LoginPage.tsx
@@ -1,12 +1,11 @@
-// import { useNavigate } from "react-router-dom"; // unused
 import {
   LogoIcon,
   KakaoIcon,
   NaverIcon,
   GoogleIcon,
-} from "../../../shared/components/icons/LoginIcons";
-import { SocialLoginButton } from "../components/SocialLoginButton";
-import loginBgImage from "../../../shared/assets/images/img_login_bg.png";
+} from "@/shared/components/icons/LoginIcons";
+import { SocialLoginButton } from "@/features/auth/components/SocialLoginButton";
+import loginBgImage from "@/shared/assets/images/img_login_bg.png";
 import {
   KAKAO_CLIENT_ID,
   KAKAO_REDIRECT_URI,
@@ -14,7 +13,7 @@ import {
   NAVER_REDIRECT_URI,
   GOOGLE_CLIENT_ID,
   GOOGLE_REDIRECT_URI,
-} from "../api/auth_api";
+} from "@/features/auth/api/auth_api";
 
 export const LoginPage = () => {
   const handleSocialLogin = (provider: string) => {
